requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
openai==1.10.0
python-dotenv==1.0.0
gradio==4.16.0
